Add files via upload
Welcome to the project repository which walks you through the process of creating a connection to a database and adding a "Person" object with its attributes including refueling details and food service usage. This project aims to provide a detailed overview of how to interact with a database through the execution of an application.

Goals
The main goal of this project is to demonstrate how to connect to a database and add details of a person, including fuel location and food service usage, by running an application. Key objectives include:

Database creation
Connecting to the Application
Definition of the Person Object
Data entry
Running the Application

Used technologies
Programming language: Java
Database: Postgresql